Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,9 +2,8 @@
 
 int main()
 {
-    int n1,n2,n3;
-
-    scanf("%d%d%d",&n1,&n2,&n3);
+    float n1,n2,n3;
+    scanf("%f%f%f",&n1,&n2,&n3);
     printf("Average: %.2f",(n1+n2+n3)/3);
     return 0;
 }
